Bug 1727191 - Generate more CacheIR infrastructure from YAML r?jandem,mgaudet
Expand `CacheIROps.yaml` into the more general `CacheIR.yaml`, which
also contains definitions for the various CacheKinds and their inputs
and outputs, and metadata for the IonICs that implement them. This lets
us autogenerate more repetitive code, such as the big switch statements
in `IonCacheIRCompiler::init` and
`CodeGenerator::visitOutOfLineICFallback`.

While we're here, we can get rid of the dynamic `stubDataOffset_` field on
`CacheIRStubInfo`. This is used to read back stub fields from
ICCacheIRStub (Baseline) and IonICStub (Ion). But when we do this
reading, we always know at compile time the type of stub we're reading,
and thus the offset of the stub fields for IC stubs of that type.

We can also avoid some frame-walking by recording fallback call
addresses when we attach Ion IC stubs, eliminating the need for
`GetReturnAddressToIonCode`. This change means that generating stubs is
no longer dependent on script runtime state, opening the door to
generating stubs in places other than OOL fallback calls.

Differential Revision: https://phabricator.services.mozilla.com/D123393